Move INCLUDE, RAISE, EXCLUDE to `compat` and `UnsupportedValueError` to `exceptions`
