+++ marshmallow_jsonschema/__init__.py
@@ -3,6 +3,7 @@ from pkg_resources import get_distribution
 __version__ = get_distribution("marshmallow-jsonschema").version
 __license__ = "MIT"
 
-from .base import JSONSchema, UnsupportedValueError
+from .base import JSONSchema
+from .exceptions import UnsupportedValueError
 
 __all__ = ("JSONSchema", "UnsupportedValueError")
