+++ marshmallow_jsonschema/exceptions.py
@@ -1,0 +1,2 @@
+class UnsupportedValueError(Exception):
+    pass
