+++ requirements-test.txt
@@ -1,5 +1,5 @@
 coverage>=4.5.3
-jsonschema>=2.6.0
+jsonschema>=3.0.1
 pytest>=4.6.3
 pytest-cov
