+++ tests/test_additional_properties.py
@@ -1,13 +1,8 @@
 import pytest
 from marshmallow import Schema, fields
 
-from marshmallow_jsonschema.base import (
-    JSONSchema,
-    UnsupportedValueError,
-    RAISE,
-    INCLUDE,
-    EXCLUDE,
-)
+from marshmallow_jsonschema import UnsupportedValueError, JSONSchema
+from marshmallow_jsonschema.compat import RAISE, INCLUDE, EXCLUDE
 from . import validate_and_dump
